:construction: fix(chore): exclude branch's job from analysis (#10741)

+++ Jenkinsfile
@@ -25,6 +25,12 @@ if (env.CHANGE_BRANCH) {
   buildBranch = env.CHANGE_BRANCH
 }
 
+// Skip sonarQ analysis on branch without PR  - Unable to merge
+def securityAnalysisRequired = 'yes'
+if (!env.CHANGE_ID && env.BUILD == 'CI') {
+    securityAnalysisRequired = 'no'
+}
+
 /*
 ** Functions
 */
@@ -136,16 +142,20 @@ try {
           trendChartType: 'NONE'
         )
 
-        // Run sonarQube analysis
-        unstash 'git-sources'
-        sh 'rm -rf centreon-web && tar xzf centreon-web-git.tar.gz'
-        withSonarQubeEnv('SonarQubeDev') {
-          sh "./centreon-build/jobs/web/${serie}/mon-web-analysis.sh"
-        }
-        timeout(time: 10, unit: 'MINUTES') {
-          def qualityGate = waitForQualityGate()
-          if (qualityGate.status != 'OK') {
-            error "Pipeline aborted due to quality gate failure: ${qualityGate.status}"
+        if (securityAnalysisRequired == 'no') {
+          Utils.markStageSkippedForConditional('sonar')
+        } else {
+          // Run sonarQube analysis
+          unstash 'git-sources'
+          sh 'rm -rf centreon-web && tar xzf centreon-web-git.tar.gz'
+          withSonarQubeEnv('SonarQubeDev') {
+            sh "./centreon-build/jobs/web/${serie}/mon-web-analysis.sh"
+          }
+          timeout(time: 10, unit: 'MINUTES') {
+            def qualityGate = waitForQualityGate()
+            if (qualityGate.status != 'OK') {
+              error "Pipeline aborted due to quality gate failure: ${qualityGate.status}"
+            }
           }
         }
       }
@@ -165,7 +175,7 @@ try {
         sh "./centreon-build/jobs/web/${serie}/mon-web-package.sh centos7"
         archiveArtifacts artifacts: "rpms-centos7.tar.gz"
         stash name: "rpms-centos7", includes: 'output/noarch/*.rpm'
-        sh 'rm -rf output'      
+        sh 'rm -rf output'
       }
     }
     //'packaging centos8': {
@@ -214,7 +224,7 @@ try {
       }
     }
   }
-  
+
   stage('Docker creation') {
     parallel 'Docker centos7': {
       node {
